Proper skippage for sanity check in B::Deparse's core.t.
The check was being skipped when PERL_CORE was not defined or if
running on VMS, but the test count was not being updated, so it
would be wrong and cause the test to fail in these cases.  So use
the standard skip facility to handle the skipping.  Plus the test
works fine on VMS so go ahead and run it there as well.