Allow cow with $magic = $hashkey
This was brought up in
<https://rt.perl.org/rt3/Ticket/Display.html?id=114820#txn-1165898>.

There is no reason we cannot assigned a shared hash key to a magical
scalar.  The only destination flag in CAN_COW_MASK that makes COW
assignment questionable is SVf_BREAK.  If such an assignment can hap-
pen (and I don’t believe it actually can), we will end up with unbal-
anced string table warnings.  So change the CAN_COW_MASK check to an
SVf_BREAK check.