Documentation of non-handle IO::Handle methods in perlvar.pod
The documentation to IO::Handle says that certain methods may not be set
on file handles, only statically. It would be nice (in my perhaps
not-so-humble opinion) if the distinction between per-filehandle and
non-perl-filehandle methods were recognized in perlvar.pod. The appended
universal diff is my cut at this.

diff a/pod/perlvar.pod b/pod/perlvar.pod