File::Spec::Unix->catfile should preserve leading dots
catfile() used to call canonpath() twice : one on the file part (its last
argument) and one on the concatenated path. But canonpath() strips any
leading dots, making it impossible to explicitely prepend a dot to indicate
the current directory. For example, this is required when you want to pass
the relative path to an executable to system().

Starting from this commit, catfile() joins all segments with a slash and
removes repeated slashes or unneeded /././ inside the resulting path.
Unfortunately, the easiest way to achieve this is to inline two
substitutions from canonpath(). On the other hand, the new catfile() is
at least 60% faster than the old one.

This fixes rt.perl.org #113898.